requirements.txt
adjustText==0.7.3
anndata==0.8.0
jupyter==1.0.0
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
python-dotenv==0.21.0
scanpy==1.9.1
scgen==2.1.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
tensorboard==2.11.0
torch==1.13.0
torchvision==0.14.0